Move pdf related files
