--- src/main/java/uk/gov/hmcts/reform/slc/services/steps/getpdf/UnableToExtractIdFromFileNameException.java
@@ -1,12 +1,0 @@
-package uk.gov.hmcts.reform.slc.services.steps.getpdf;
-
-public class UnableToExtractIdFromFileNameException extends RuntimeException {
-
-    public UnableToExtractIdFromFileNameException(Exception inner) {
-        super(inner);
-    }
-
-    public UnableToExtractIdFromFileNameException(String message) {
-        super(message);
-    }
-}
